Update aem.weight.edges.R
I just got a comment about an odd behaviour of the aem.weight.edges function; it sometimes generated negative weights. This is a behaviour that should not happen, weights should always be positive. The reason for this behaviour is that the maximum distance between two nearby site (`max(links.length)`) is larger than `max.d`. I added a trap in the code to send an error message when `max(links.length) > max.d`.