Add epsilon to threshold value
We observe instability in results (different results for Pierre and Daniel). This is related to the computation of the nb object by dnearneigh and the bug is observed when several distances are equal to the threshold value (e.g. regular sampling). We now add an epsilon for all cases (only the distance case before) to solve the problem and avoid equality between distances and threshold value